# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  description: |
    Phase 1 backend API — auth, restaurant leads, durable city scraping, OCR/review gates,
    demo sites, and **email outreach campaigns**.

    This document covers the Phase 1 lead-to-outreach and reservation surface. The router
    remains the source of truth for separate company-consultation and legacy public asset
    routes that are outside this workflow specification.

    ## Authentication
    Protected routes use `Authorization: Bearer <access_token>` from signup or login.

    ## Roles
    - `internal_admin` — full restaurant admin (seed via `make seed-admin`)
    - `restaurant_owner` — scoped to assigned restaurants only
    - `developer` — health endpoints (local/test signup only)

    ## Lead preparation and outreach flow (admin)
    1. `POST /api/v1/scrape-jobs` — create or return the active durable city/niche scrape job.
    2. The scrape worker uses Google Places first and Apollo only for missing contact details.
       It pauses at 500 combined provider requests and resumes after its durable `resume_at`.
    3. OCR moves each profile through `pending`, `running`, then `verified`, `no_images`, or
       `failed`. Only `verified` can advance to outreach.
    4. OCR verification automatically creates or refreshes draft demo/campaign artifacts.
    5. Inspect the protected profile and demo review-preview endpoints; neither discloses
       the demo bearer token.
...

...

...

    ## Local email testing
...
servers:
...
paths:
...
  /api/v1/restaurants:
    get:
      tags: [Restaurants]
      summary: List accessible restaurants
      operationId: listRestaurants
      security:
        - bearerAuth: []
      description: |
        - `internal_admin` — all restaurants (subject to filters)
        - `restaurant_owner` — only restaurants where user is a member
        Archived restaurants are hidden unless `include_archived=true`.
      parameters:
        - $ref: '#/components/parameters/restaurantNameFilter'
        - $ref: '#/components/parameters/statusFilter'
        - $ref: '#/components/parameters/isContactedFilter'
        - $ref: '#/components/parameters/shownInterestFilter'
        - $ref: '#/components/parameters/ocrStatusFilter'
        - $ref: '#/components/parameters/includeArchivedFilter'
      responses:
        '200':
          description: Restaurant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Restaurants]
      summary: Create restaurant lead
      operationId: createRestaurant
      security:
        - bearerAuth: []
      description: Requires role `internal_admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestaurantRequest'
            example:
              name: PM Demo Thai Kitchen
              email: lead@pmdemo.test
      responses:
        '201':
          description: Restaurant created (status defaults to `lead`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/restaurants/{id}:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    get:
      tags: [Restaurants]
      summary: Get restaurant by ID
      operationId: getRestaurant
      security:
        - bearerAuth: []
      description: |
        Admin or restaurant owner with membership.
        Archived restaurants return 404 for non-admin users.
      responses:
        '200':
          description: Restaurant record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Restaurants]
      summary: Update restaurant fields
      operationId: updateRestaurant
      security:
        - bearerAuth: []
      description: |
        Requires role `internal_admin`. The admin UI treats `is_contacted` and
        `shown_interest` as read-only automation results: confirmed provider sends set
        contacted/emailed, while tracked link clicks set shown-interest/interested.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRestaurantRequest'
            example:
              is_contacted: true
      responses:
        '200':
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Restaurants]
      summary: Archive restaurant (soft delete)
      operationId: archiveRestaurant
      security:
        - bearerAuth: []
      description: Sets `status` to `archived`. Requires role `internal_admin`.
      responses:
        '200':
          description: Archived restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from signup or login
  parameters:
    restaurantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    restaurantNameFilter:
      name: restaurant
      in: query
      schema:
        type: string
      description: Case-insensitive substring match on restaurant name
    statusFilter:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/RestaurantStatus'
    isContactedFilter:
      name: is_contacted
      in: query
      schema:
        type: boolean
    shownInterestFilter:
      name: shown_interest
      in: query
      schema:
        type: boolean
    ocrStatusFilter:
      name: ocr_status
      in: query
      schema:
        $ref: '#/components/schemas/OCRStatus'
      description: Filter restaurant rows by profile OCR status; use `verified` for verified-only.
    includeArchivedFilter:
      name: include_archived
      in: query
      schema:
        type: boolean
        default: false
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Authentication required.
    Forbidden:
      description: Insufficient role or restaurant access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: You do not have access to this restaurant.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: Restaurant was not found.
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: invalid_request
            message:
              type: string
              example: Request body must be valid JSON.
...
    RestaurantStatus:
      type: string
      enum:
        - lead
        - demo_ready
        - emailed
        - interested
        - client_onboarding
        - active_client
        - lost
        - archived
    Restaurant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        ocr_status:
          $ref: '#/components/schemas/OCRStatus'
        status:
          $ref: '#/components/schemas/RestaurantStatus'
        is_contacted:
          type: boolean
        shown_interest:
          type: boolean
        email_sent:
          type: boolean
          description: True only after a confirmed provider-accepted outreach send.
        email_send_count:
          type: integer
          minimum: 0
          description: Number of confirmed sends recorded for this restaurant.
        last_email_sent_at:
          type: string
          format: date-time
          nullable: true
        last_email_send_sequence:
          type: integer
          format: int64
          nullable: true
          description: Global delivery-attempt sequence associated with the latest confirmed send.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RestaurantList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Restaurant'
    CreateRestaurantRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    UpdateRestaurantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        is_contacted:
          type: boolean
        shown_interest:
          type: boolean
...
    OCRStatus:
      type: string
      enum: [pending, running, verified, no_images, failed]
      description: |
        `verified` means at least one usable image was processed successfully and is the
        only OCR state eligible for profile approval. `no_images` is not equivalent to verified.
...